Changed type of MinMax from int*int to long*long.

+++ core/src/main/scala/com/jabong/dap/data/acq/common/MinMax.scala
@@ -3,4 +3,4 @@ package com.jabong.dap.data.acq.common
 /**
  * Created by Abhay on 9/6/15.
  */
-case class MinMax(min: Int, max: Int)
+case class MinMax(min: Long , max: Long)
